+++ HISTORY.rst
@@ -1,7 +1,7 @@
 History
 =======
 
-1.3.0 (unreleased)
+1.3.0 (2019-03-07)
 ------------------
 
 - Support names renamed with as during import.